Increased the value of "q" and "p" to the maximum value of int (maximum value: 2 147 483 647)

+++ TextRSA.java
@@ -1,10 +1,12 @@
 package RSA;
 
+import java.math.BigInteger;
+
 public class TextRSA extends IntRSA {
     String encryptedText;
     String decryptedText;
     String textMessage;
-    public TextRSA(int p, int q){
+    public TextRSA(BigInteger p, BigInteger q){
         super(p, q);
     }
 
@@ -42,4 +44,5 @@ public class TextRSA extends IntRSA {
         return decryptedText;
     }
 }
+
